# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal.Schedule:
    properties:
      created_at:
        type: string
      ends_at:
        type: string
      hall_id:
        type: integer
      id:
        type: integer
      movie_id:
        type: integer
      price:
        type: number
      starts_at:
        type: string
      version:
        type: integer
    type: object
...
  internal.Ticket:
    properties:
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      schedule_id:
        type: integer
      seat_id:
        type: integer
      state_changed_at:
        type: string
      state_id:
        $ref: '#/definitions/internal.TicketState'
      version:
        type: integer
    type: object
  internal.TicketState:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - TicketStateUnsold
    - TicketStateLocked
    - TicketStateSold
...
  main.CreateScheduleResponse:
    properties:
      schedule:
        $ref: '#/definitions/internal.Schedule'
    type: object
...
  main.LockTicketResponse:
    properties:
      ticket:
        $ref: '#/definitions/internal.Ticket'
    type: object
  main.ResponseError:
    properties:
      error:
        description: Error
        type: string
    type: object
  main.ResponseMessage:
    properties:
      message:
        description: Message
        type: string
    type: object
...
  main.UpdateScheduleResponse:
    properties:
      schedule:
        $ref: '#/definitions/internal.Schedule'
    type: object
  main.ViolationsMessage:
    properties:
      errors:
        additionalProperties:
          type: string
        description: errors
        type: object
    type: object
host: https://localhost:8080
info:
  contact:
    email: ahamerdev@gmail.com
    name: Ahmed Amer
  description: a simple movie reservation system api for educational purposes
  title: Movie Reservation System API
  version: "1.0"
paths:
...
  /schedules:
    get:
      consumes:
      - application/json
      description: Gets a list of schedules by search paramaters
      parameters:
      - description: movie_id
        in: query
        name: movie_id
        required: true
        type: integer
      - description: hall_id
        in: query
        name: hall_id
        required: true
        type: integer
      - description: page number
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: page_size
        required: true
        type: integer
      - description: sort paramterers (id, price, starts_at, ends_at) prefix with
          - to sort descending
        in: query
        name: sort
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.CreateScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Gets a list of schedules
      tags:
      - schedules
    post:
      consumes:
      - application/json
      description: creates a schedule for a given movie and hall
      parameters:
      - description: movie_id
        in: body
        name: movie_id
        required: true
        schema:
          type: integer
      - description: hall_id
        in: body
        name: hall_id
        required: true
        schema:
          type: integer
      - description: price
        in: body
        name: price
        required: true
        schema:
          type: string
      - description: starts at
        in: body
        name: starts_at
        required: true
        schema:
          type: string
      - description: ends at
        in: body
        name: ends_at
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.CreateScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a schedule
      tags:
      - schedules
  /schedules/{id}:
    delete:
      consumes:
      - application/json
      description: deletes a schedule by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UpdateScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Deletes a schedule
      tags:
      - schedules
    put:
      consumes:
      - application/json
      description: updates a schedule by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: price
        in: body
        name: price
        required: true
        schema:
          type: string
      - description: starts at
        in: body
        name: starts_at
        required: true
        schema:
          type: string
      - description: ends at
        in: body
        name: ends_at
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UpdateScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Updates a schedule
      tags:
      - schedules
  /schedules/{id}/tickets:
    get:
      consumes:
      - application/json
      description: gets a list of tickets for a given schedule
      parameters:
      - description: schedule id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Gets a list of tickets
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: creates the tickets for a given schedule
      parameters:
      - description: schedule id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates the tickets
      tags:
      - tickets
...
  /tickets/{id}/lock:
    post:
      consumes:
      - application/json
      description: locks a ticket to a given user for some time
      parameters:
      - description: ticket id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.LockTicketResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Locks a ticket
      tags:
      - tickets
  /tickets/{id}/unlock:
    post:
      consumes:
      - application/json
      description: unlocks a ticket
      parameters:
      - description: ticket id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.LockTicketResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Unlocks a ticket
      tags:
      - tickets
...
swagger: "2.0"